Vodafone-Project-main/requirements.txt
absl-py==1.0.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cmake==3.22.4
colorama==0.4.4
comtypes==1.1.11
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dlib==19.23.1
entrypoints==0.4
executing==0.8.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.15.3
Flask==2.1.1
flatbuffers==2.0
fonttools==4.32.0
future==0.18.2
gast==0.5.3
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
imutils==0.5.4
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyterlab-pygments==0.2.1
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
playsound==1.2.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
pypiwin32==223
pyrsistent==0.18.1
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
ratelim==0.1.6
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2
stack-data==0.2.0
streamlit==1.8.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.6.0
wrapt==1.14.0
zipp==3.8.0
